+++ src/Objects/ObjectStaticBox.c
@@ -7,7 +7,7 @@
 int ObjectStaticBoxInit(Object *obj, Vec2F position) {
 	PROPAGATE_ERROR(ObjectInit(obj, position));
 	
-	uint32_t phyId = 0;
+	uint64_t phyId = 0;
 	PhysicsComponent* phy = ObjectAddAndInitPhysics(obj, &phyId);
 	phy->body = Box2DUtilsCreateStaticBox(phyId, position, STATIC_OBJECT_CATEGORY, ((Vec2F) { 0.875, 0.125 }));
